follow-up tp [r4845] speedup for ACCEPT date + time
libcob:
* common.c (cob_accept_date, cob_accept_date_yyyymmdd, cob_accept_day, cob_accept_day_yyyyddd, cob_accept_time, cob_accept_microsecond_time): use binary intermediate field instead of numeric edited, replacing "expensive" sprintf calls and multiple de-editing/editing to max. one minimal editing (via move.c)
* common.c (cob_move_to_group_as_alnum): new function to handle MOVE to group as MOVE to alphanumeric field, used in functions above
* common.c (cob_display_arg_number, cob_accept_arg_number, cob_cmp): minor refactoring to use COB_FIELD_INIT
* common.c (check_current_date, cob_get_current_datetime), coblocal.h (cob_settings->cob_time_constant_is_calculated): skip expensive call to localtime + mktime if COB_CURRENT_DATE is set and "complete", by doing it already for the constant

+++ libcob/call.c
@@ -2074,7 +2074,7 @@ const char *
 cob_get_field_str_buffered (const cob_field *f)
 {
 	char	*buff = NULL;
-	size_t	size = cob_get_field_size (f) + 1;
+	size_t	size = (size_t)cob_get_field_size (f) + 1;
 
 	if (size > 0) {
 		if (size < 32) {
